Final commit fro #30

+++ setup/puppet/modules/wordpress/manifests/install-wp.pp
@@ -1,6 +1,6 @@
 define wordpress::install-wp($wp_remote_location, $mode = 0644, $wp_localpath, $metod = 'GIT', $module_path, $wp_db_prefix, $wp_subdomain,
-		    $wp_admin_email, $web_owner, $web_group, $wp_dbhost, $wp_dbname, $wp_dbuser, $wp_dbpass, $apache_conf, 
-		    $wp_admin_user, $wp_admin_password, $wp_title, $wp_url, $wp_mysql_port, $wp_plugin_MU, $wp_apache_localpath ){
+		    $wp_admin_email, $web_owner, $web_group, $wp_dbhost, $wp_dbname, $wp_dbuser, $wp_dbpass, $apache_conf, $wp_config_path,
+		    $wp_admin_user, $wp_admin_password, $wp_title, $wp_url, $wp_mysql_port, $wp_plugin_MU, $wp_apache_localpath, $wp_env ){
 #    include apache
     include extra
     
@@ -23,26 +23,39 @@ define wordpress::install-wp($wp_remote_location, $mode = 0644, $wp_localpath, $
       
       exec{"wp-cli-install":
         command => "/bin/cp $module_path/wordpress/files/wpcli /tmp/ \
-		&& if [ $wp_subdomain == \"Yes\" ]; then /usr/bin/php /tmp/wpcli core multisite-install --subdomains --path=$wp_localpath --url=$wp_url --title=\"$wp_title\" --admin_user=$wp_admin_user --admin_password=$wp_admin_password --admin_email=$wp_admin_email; \
-		else /usr/bin/php /tmp/wpcli core multisite-install --path=$wp_localpath --url=$wp_url --title=\"$wp_title\" --admin_user=$wp_admin_user --admin_password=$wp_admin_password --admin_email=$wp_admin_email; fi \
-		&& if [ $wp_plugin_MU == \"Yes\" ]; then /usr/bin/php /tmp/wpcli plugin install \"WordPress MU Domain Mapping\" --path=$wp_localpath; \
-		/bin/mv $wp_localpath/wp-content/plugins/wordpress-mu-domain-mapping/sunrise.php $wp_localpath/wp-content/; \
-		/bin/sed -i \"/define('ABSPATH', dirname(__FILE__) . '\\/');/a define( 'SUNRISE', 'on' );\" $wp_localpath/wp-config.php; fi \
-		&& /bin/chown -R $web_owner.$web_group $wp_localpath \
-		&& rm -fr /tmp/wpcli",
+    	&& if [ -f $wp_config_path/$wp_env-wp-config.php ]; then /bin/sed -i \"s/define( 'SUNRISE', 'on' );//g\" $wp_config_path/$wp_env-wp-config.php \
+    		 && mv $wp_config_path\\/$wp_env-wp-config.php $wp_localpath/wp-config.php; \
+		 if [ $wp_subdomain == \"Yes\" ]; then /bin/cp $module_path/wordpress/templates/htaccesSubdomain.erb $wp_localpath/.htaccess; \
+		 /usr/bin/php /tmp/wpcli core multisite-install --subdomains --path=$wp_localpath --url=$wp_url --title=\"$wp_title\" --admin_user=$wp_admin_user --admin_password=$wp_admin_password --admin_email=$wp_admin_email; \
+		 else /bin/cp $module_path/wordpress/templates/htaccesSubfolder.erb $wp_localpath/.htaccess; \
+		 /usr/bin/php /tmp/wpcli core multisite-install --path=$wp_localpath --url=$wp_url --title=\"$wp_title\" --admin_user=$wp_admin_user --admin_password=$wp_admin_password --admin_email=$wp_admin_email; fi; \
+		 if [ $wp_plugin_MU == \"Yes\" ]; then /usr/bin/php /tmp/wpcli plugin install \"WordPress MU Domain Mapping\" --path=$wp_localpath; \
+		 /bin/mv $wp_localpath/wp-content/plugins/wordpress-mu-domain-mapping/sunrise.php $wp_localpath/wp-content/; \
+		 /bin/sed -i \"/define('ABSPATH', dirname(__FILE__) . '\\/');/a define( 'SUNRISE', 'on' );\" $wp_localpath/wp-config.php; fi; \
+	else rm -f $wp_localpath/wp-config.php; \
+		 /usr/bin/php /tmp/wpcli core config --path=$wp_localpath --dbname=$wp_dbname --dbuser=$wp_dbuser --dbpass=$wp_dbpass --dbhost=$wp_dbhost:$wp_mysql_port --dbprefix=$wp_db_prefix; \
+		 /usr/bin/php /tmp/wpcli db reset --path=$wp_localpath --yes; \
+		 if [ $wp_subdomain == \"Yes\" ]; then /bin/cp $module_path/wordpress/templates/htaccesSubdomain.erb $wp_localpath/.htaccess; \
+		 /usr/bin/php /tmp/wpcli core multisite-install --subdomains --path=$wp_localpath --url=$wp_url --title=\"$wp_title\" --admin_user=$wp_admin_user --admin_password=$wp_admin_password --admin_email=$wp_admin_email; \
+		 else /bin/cp $module_path/wordpress/templates/htaccesSubfolder.erb $wp_localpath/.htaccess; \
+		 /usr/bin/php /tmp/wpcli core multisite-install --path=$wp_localpath --url=$wp_url --title=\"$wp_title\" --admin_user=$wp_admin_user --admin_password=$wp_admin_password --admin_email=$wp_admin_email; fi; \
+		 if [ $wp_plugin_MU == \"Yes\" ]; then /usr/bin/php /tmp/wpcli plugin install \"WordPress MU Domain Mapping\" --path=$wp_localpath; \
+		 /bin/mv $wp_localpath/wp-content/plugins/wordpress-mu-domain-mapping/sunrise.php $wp_localpath/wp-content/; \
+		 /bin/sed -i \"/define('ABSPATH', dirname(__FILE__) . '\\/');/a define( 'SUNRISE', 'on' );\" $wp_localpath/wp-config.php; fi; \
+	    fi \
+    	&& /bin/mv $wp_localpath/wp-config.php $wp_config_path\\/$wp_env-wp-config.php \
+    	&& /bin/cp $module_path/wordpress/templates/wp-config.erb $wp_localpath/wp-config.php \
+	&& /bin/chown -R $web_owner.$web_group $wp_localpath \
+	&& rm -fr /tmp/wpcli",
 	path => "/bin:/usr/bin",
+#	onlyif => "[ -d $wp_config_path -a -z \"$(ls $wp_config_path\\/$wp_env-wp-config.php)\" ]",
 	require => Exec["get_web_${tmppath}"],
       }
       
   exec{"get_web_${$tmppath}":
   
 	    command => "$command \
-			&& /bin/cp $module_path/wordpress/files/wpcli /tmp/ \
 			&& /bin/tar -zxf $tmppath -C $wp_localpath --strip-components 1 \
-			&& /usr/bin/php /tmp/wpcli core config --path=$wp_localpath --dbname=$wp_dbname --dbuser=$wp_dbuser --dbpass=$wp_dbpass --dbhost=$wp_dbhost:$wp_mysql_port --dbprefix=$wp_db_prefix \
-			&& if [ $wp_subdomain == \"Yes\" ]; then /bin/cp $module_path/wordpress/templates/htaccesSubdomain.erb $wp_localpath/.htaccess; \
-			else /bin/cp $module_path/wordpress/templates/htaccesSubfolder.erb $wp_localpath/.htaccess; fi \
-			&& rm -fr /tmp/wpcli \
 			&& /bin/rm -f $tmppath",
 	    require => Notify[ note-wp-download ],
 #	    subscribe => File["$tmppath"],
@@ -50,6 +63,7 @@ define wordpress::install-wp($wp_remote_location, $mode = 0644, $wp_localpath, $
 #	    refreshonly => true,
 	    path => "/bin:/usr/bin",
 	    onlyif => "[ -d $wp_localpath -a -z \"$(ls -A -I '.gitkeep' $wp_localpath)\" ]",
+#	    onlyif => "[ -d $wp_localpath -a -z \"$(ls $wp_localpath\\/wp-config.php)\" ]",
     }
 
  file{"${wp_localpath}":
@@ -85,9 +99,11 @@ define wordpress::install-wp($wp_remote_location, $mode = 0644, $wp_localpath, $
 		&& if [ $wp_plugin_MU == \"Yes\" ]; then /usr/bin/php /tmp/wpcli plugin install \"WordPress MU Domain Mapping\" --path=$wp_localpath; \
 		/bin/mv $wp_localpath/wp-content/plugins/wordpress-mu-domain-mapping/sunrise.php $wp_localpath/wp-content/; \
 		/bin/sed -i \"/define('ABSPATH', dirname(__FILE__) . '\\/');/a define( 'SUNRISE', 'on' );\" $wp_localpath/wp-config.php; fi \
+		&& /bin/cp $module_path/wordpress/templates/wp-config.erb $wp_localpath/wp-config.php \
 		&& /bin/chown -R $web_owner.$web_group $wp_localpath \
 		&& rm -fr /tmp/wpcli",
 	path => "/bin:/usr/bin",
+	onlyif => "[ -d $wp_config_path -a -z \"$(ls $wp_config_path\\/$wp_env-wp-config.php)\" ]",
 	require => Exec["get_git_${wp_localpath}"],
       }
